+++ ving/docs/change-log.md
@@ -7,6 +7,7 @@ outline: deep
 
 ### 2024-06-25
 * Disabled nuxt devtools, because I never use them and they are spewing junk into the console.
+* Workaround: calling useVingKind().all() when doing an SSR without a middleware causes Nuxt to crash #168
 
 ### 2024-06-24
 * Fixed: pager appears to be broken #167